+++ SpringData/Api/src/main/java/com/example/api/ProductDTO.java
@@ -4,6 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.swing.plaf.basic.BasicDesktopIconUI;
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,7 +14,7 @@ public class ProductDTO {
 
     private Long id;
     private String title;
-    private Long cost;
+    private BigDecimal cost;
 
 
 }
