+++ SpringData/cart-service/src/main/java/com/example/cart/model/CartItem.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,6 +13,6 @@ public class CartItem {
     private Long productId;
     private String productTitle;
     private int quantity;
-    private Long costPerProduct;
-    private Long cost;
+    private BigDecimal costPerProduct;
+    private BigDecimal cost;
 }
